requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile requirements.in
#
biopython==1.79
    # via -r requirements.in
cycler==0.10.0
    # via matplotlib
kiwisolver==1.3.2
    # via matplotlib
matplotlib==3.4.3
    # via
    #   -r requirements.in
    #   seaborn
numpy==1.21.2
    # via
    #   -r requirements.in
    #   biopython
    #   matplotlib
    #   pandas
    #   scipy
    #   seaborn
pandas==1.3.3
    # via
    #   -r requirements.in
    #   seaborn
pillow==8.4.0
    # via matplotlib
pyparsing==2.4.7
    # via matplotlib
python-dateutil==2.8.2
    # via
    #   matplotlib
    #   pandas
pytz==2021.3
    # via pandas
scipy==1.7.1
    # via
    #   -r requirements.in
    #   seaborn
seaborn==0.11.2
    # via -r requirements.in
six==1.16.0
    # via
    #   cycler
    #   python-dateutil
